+++ Lunar/DataModels/ReportPostResponseModel.swift
@@ -7,16 +7,20 @@
 
 import Foundation
 
-struct ReportPostResponseModel: Codable {
-  let postReportView: ReportModel
+struct ReportResponseModel: Codable {
+  let postReportModel: PostReportModel?
+  let commentReportModel: CommentReportModel?
+  let privateMessageReportModel: PrivateMessageReportModel?
 
   enum CodingKeys: String, CodingKey {
-    case postReportView = "post_report_view"
+    case postReportModel = "post_report_view"
+    case commentReportModel = "comment_report_view"
+    case privateMessageReportModel = "private_message_report_view"
   }
 }
 
-struct ReportModel: Codable {
-  let postReport: Report
+struct PostReportModel: Codable {
+  let postReport: PostReportObject
   let post: Post
   let community: Community
   let creator, postCreator: Person
@@ -32,7 +36,38 @@ struct ReportModel: Codable {
   }
 }
 
-struct Report: Codable {
+struct CommentReportModel: Codable {
+  let commentReport: CommentReportObject
+  let comment: Comment
+  let post: Post
+  let community: Community
+  let creator, commentCreator: Person
+  let counts: Counts
+  let creatorBannedFromCommunity: Bool
+
+  enum CodingKeys: String, CodingKey {
+    case commentReport = "comment_report"
+    case comment, post, community, creator
+    case commentCreator = "comment_creator"
+    case counts
+    case creatorBannedFromCommunity = "creator_banned_from_community"
+  }
+}
+
+struct PrivateMessageReportModel: Codable {
+  let privateMessageReport: PrivateMessageReportObject
+  let privateMessage: PrivateMessage
+  let privateMessageCreator, creator: Person
+
+  enum CodingKeys: String, CodingKey {
+    case privateMessageReport = "private_message_report"
+    case privateMessage = "private_message"
+    case privateMessageCreator = "private_message_creator"
+    case creator
+  }
+}
+
+struct PostReportObject: Codable {
   let id, creatorID, postID: Int
   let originalPostName: String
   let originalPostURL: String
@@ -47,6 +82,36 @@ struct Report: Codable {
     case originalPostName = "original_post_name"
     case originalPostURL = "original_post_url"
     case originalPostBody = "original_post_body"
+    case reason, resolved, published
+  }
+}
+
+struct CommentReportObject: Codable {
+  let id, creatorID, commentID: Int
+  let originalCommentText, reason: String
+  let resolved: Bool
+  let published: String
+
+  enum CodingKeys: String, CodingKey {
+    case id
+    case creatorID = "creator_id"
+    case commentID = "comment_id"
+    case originalCommentText = "original_comment_text"
+    case reason, resolved, published
+  }
+}
+
+struct PrivateMessageReportObject: Codable {
+  let id, creatorID, privateMessageID: Int
+  let originalPmText, reason: String
+  let resolved: Bool
+  let published: String
+
+  enum CodingKeys: String, CodingKey {
+    case id
+    case creatorID = "creator_id"
+    case privateMessageID = "private_message_id"
+    case originalPmText = "original_pm_text"
     case reason, resolved, published
   }
 }